+++ views/userview2-2.handlebars
@@ -2,24 +2,31 @@
   <h2 class="form-h2">Here is your current reservation</h2>
   <session class="restaurant-profile" id="restaurant-profile">
     <img class="restaurant-pic" src="js/Restaurant-profile-pic.jpeg" />
-    <h3>{{Owner.restaurant_name}}</h3>
+    <h3 id="restaurant-name">{{Owner.restaurant_name}}</h3>
   </session>
   <div class="reservation-form" id="reservation-form">
     <p>Customer Name:</p>
-    <p><strong>{{Customer.name}}</strong></p>
+    <p name = "name"><strong>{{Customer.name}}</strong></p>
     <p>Customer Email:</p>
-    <p><strong>{{Customer.email}}</strong></p>
+    <p name="email"><strong>{{Customer.email}}</strong></p>
     <p>Party Size:</p>
-    <p><strong>{{party_size}}</strong></p>
+    <p name="party_size"><strong>{{party_size}}</strong></p>
     <p>Date:</p>
-    <p><strong>{{reservation_date}}</strong></p>
+    <p name="date"><strong>{{reservation_date}}</strong></p>
     <p>Time:</p>
-    <p><strong>{{reservation_time}}</strong></p>
+    <p name="time"><strong>{{reservation_time}}</strong></p>
     <p></p>
   </div>
+  <p>{{msg}}</p>
   <button class="form-btn  " id="send-reservation">Send Email</button>
   <button class="form-btn  " id="cancle-reservation">Cancle the Reservation</button>
 
 </div>
 
 <script src="/js/userview2-2.js"></script>
+
+<h3>Dear ${req.body.name}</h3>
+<br>
+<p>You have a reservation with ${req.body.restaurant} on ${req.boy.date}</p>
+<p>at ${req.body.time}.</p>
+<p>Pary size is ${req.body.party_size}</p>
