+++ views/userview3.handlebars
@@ -1,8 +1,8 @@
 <div class="userview-form" id="reservation-confirmed-form">
 
   <h3>Your reservation has benn confirmed.</h3>
-  <h3>An email has been sent to the following email address:{{email}}</h3>
-
+  <h3>Would you like an email of your reservaion sent to the following email address:{{email}}</h3>
+  <button class="form-btn" id="sent-email-btn">Send Email</button>
   <button class="form-btn" id="check-reservation-btn">Check Reservation</button>
   <button class="form-btn" id="cancle-btn">Cancle</button>
   <button class="form-btn" id="menu-btn">Check Menu</button>
